fix: my platform playlist show error
