+++ src/views/HotPlaylists.vue
@@ -24,24 +24,7 @@
     </div>
     <div id="hotplaylist" class="site-wrapper-innerd">
       <div id="playlist-content" class="cover-container">
-        <ul class="playlist-covers m-0 py-0 flex relative flex-wrap">
-          <li v-for="i in result" :key="i.id" class="group w-1/3 md:w-1/5 min-h-40 px-4 sm:w-1/4">
-            <div class="u-cover flex relative">
-              <img :src="i.cover_img_url" @click="showPlaylist(i.id)" class="m-auto border border-default cursor-pointer w-full object-cover" />
-              <div
-                class="bottom opacity-0 group-hover:opacity-100 cursor-pointer absolute w-8 h-8 bottom-3 right-3 ease-linear duration-200"
-                @click="directplaylist(i.id)">
-                <vue-feather type="play-circle" size="2rem"></vue-feather>
-              </div>
-            </div>
-            <div class="desc cursor-pointer">
-              <span class="title flex min-h-8" @click="showPlaylist(i.id)">{{ i.title }}</span>
-            </div>
-          </li>
-          <!-- <div class="loading_bottom">
-                              <img src="images/loading-1.gif" height="40px" />
-          </div>-->
-        </ul>
+        <PlaylistGrid :playlists="result"></PlaylistGrid>
       </div>
     </div>
   </div>
@@ -49,16 +32,12 @@
 <script setup>
 /* eslint-disable @typescript-eslint/no-unused-vars */
 
-import { l1Player } from '../services/l1_player';
 import { computed, onMounted, onUnmounted } from 'vue';
-import { useI18n } from 'vue-i18n';
-import { useRouter } from 'vue-router';
 import MediaService from '../services/MediaService';
 import EventService from '../services/EventService';
 import SourceTab from '../components/SourceTab.vue';
+import PlaylistGrid from '../components/PlaylistGrid.vue';
 
-const { t } = useI18n();
-const router = useRouter();
 let currentFilterId = $ref('');
 let result = $ref([]);
 let tab = $ref(MediaService.getSourceList()[0].name);
@@ -112,14 +91,6 @@ const changeFilter = (filterId) => {
 };
 const toggleMorePlaylists = () => {
   showMore = !showMore;
-};
-const showPlaylist = (playlistId) => {
-  router.push('/playlist/' + playlistId);
-};
-const directplaylist = async (list_id) => {
-  const data = await MediaService.getPlaylist(list_id);
-  const songs = data.tracks;
-  l1Player.playTracks(songs);
 };
 const sourceList = computed(() => MediaService.getSourceList());
 </script>
